Create model.py - move common data type definitions there
